messages/requirements.txt
django
hazelcast-python-client
django-daemon-command
python-consul